Move index offloading into a dedicated module
